+++ src/main/java/com/dez/predesign/repository/ProductRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepo extends CrudRepository<Product,Integer> {
+public interface ProductRepo extends CrudRepository<Product,Long> {
     Page<Product> findByBrand(String tag, Pageable pageable);
 
     Page<Product> findAll(Pageable pageable);
